Introduce frequentist test superclass to reduce code duplication
